FutureBuilder is implemented succesfully, UserHome and UserFeeds
The My Home page (or UserHomeActivity) is functional now.
The FutureBuilder is fixed, along with the image encoding and decoding.
The UserHomeActivity shows a grid of photos that the user has posted themselves.
The UserFeedsActivity is also functional. It fetches data / photos of other users and Displays them in a list. Like a set of cards.
Also made changes to the Constants Class. Changed some names of variables as the ones used before is restrictive (in their naming). Now all keys used for upload is uXYZ (XYZ is keyname) and for download it is dXYZ.
New class FeedItem which is a place holder for the Feeds.

Todo:
1. Add DetailView page
2. Add more details to user home page.
3. Add Like system
4. Add comment system.